+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ v1.2.2
 
 - Fix error comparing screenshots in mobile tests
 - Fix image size when enlarging a vertical image in visual testing reports
+- Move js and css out of visual html report to avoid CSP errors
 
 v1.2.1
 ------